refactor: activity certificate
