+++ packages/web/src/features/activity-certificate/components/ManageActivityCertificate.tsx
