+++ packages/web/src/features/activity-certificate/components/MyActivityCertificateProgress.tsx
