+++ packages/web/src/features/manage-club/frames/ServiceManageFrame.tsx
@@ -3,7 +3,7 @@ import React from "react";
 import FlexWrapper from "@sparcs-clubs/web/common/components/FlexWrapper";
 import FoldableSectionTitle from "@sparcs-clubs/web/common/components/FoldableSectionTitle";
 import MoreDetailTitle from "@sparcs-clubs/web/common/components/MoreDetailTitle";
-import ActivityCertificateTable from "@sparcs-clubs/web/features/manage-club/components/ActivityCertificateTable";
+import ActivityCertificateTable from "@sparcs-clubs/web/features/activity-certificate/components/ActivityCertificateTable";
 import CommonSpaceTable from "@sparcs-clubs/web/features/manage-club/components/CommonSpaceTable";
 import PrintingTable from "@sparcs-clubs/web/features/manage-club/components/PrintingTable";
 import RentalTable from "@sparcs-clubs/web/features/manage-club/components/RentalTable";
